# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CommentCreate": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "content"
        ],
        "title": "CommentCreate",
        "type": "object"
      },
      "CommentOut": {
        "description": "Something said about a post, with whatever was said back.\n\nReplies are not paged. A conversation arrives whole or the page boundary\nwould fall through the middle of one, and a thread on this app is a handful\nof lines rather than a hundred.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "replies": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReplyOut"
            },
            "title": "Replies",
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "content",
          "created_at",
          "post_id",
          "user_id",
          "user"
        ],
        "title": "CommentOut",
        "type": "object"
      },
      "CommentPage": {
        "description": "A page of comments. Deliberately the same shape as ``PostPage`` — a\nclient that can page through one can page through the other without\nlearning a second set of field names.",
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "has_prev": {
            "title": "Has Prev",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_replies": {
            "default": 0,
            "title": "Total Replies",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages",
          "has_next",
          "has_prev"
        ],
        "title": "CommentPage",
        "type": "object"
      },
      "Detail": {
        "description": "FastAPI's error body, written down so it can be named in `responses=`.\n\nEvery non-validation error this API returns is one of these. Declaring it\nis what turns \"401\" in the docs from a bare number into a shape a client\ncan code against.",
        "example": {
          "detail": "Could not validate credentials"
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Detail",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplyOut": {
        "description": "An answer to a comment.\n\n**Note what it hasn't got: replies of its own.** That is the one-level rule,\nwritten where a reader of the contract will see it rather than buried in a\nhandler — the shape of the response is the specification, and this shape\ncannot express a third level.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "parent_id": {
            "title": "Parent Id",
            "type": "integer"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "content",
          "created_at",
          "post_id",
          "user_id",
          "user",
          "parent_id"
        ],
        "title": "ReplyOut",
        "type": "object"
      },
...
      "UserOut": {
        "description": "A person, as everybody else sees them. No email: that's a credential,\nand handing it to anyone who can walk a list of ids is how address books\nget scraped.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "created_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Source on GitHub",
      "url": "https://github.com/AndrewTechTips/social-feed-application"
    },
    "description": "The backend behind **Commons** — a small public feed.\n\nPosts, comments, upvotes and accounts, with a two-token auth flow. Reading is public; writing needs an account.\n\nThe published demo at [andrewtechtips.github.io/social-feed-application](https://andrewtechtips.github.io/social-feed-application/) reimplements this same contract in the browser, because a static host has nowhere to run FastAPI. This document is the contract both of them answer to.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/AndrewTechTips/social-feed-application/blob/main/LICENSE"
    },
    "title": "Social Feed API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/comments/{id}": {
      "delete": {
        "description": "Yours to remove, and nobody else's.\n\nNote what this deliberately isn't: the author of the *post* can't delete\ncomments on it. Moderation is a feature with a scope of its own — who may\nremove what, whether the person who wrote it is told, whether it leaves a\ntombstone — and quietly granting it to post authors here would be a policy\ndecision smuggled in as an ownership check.",
        "operationId": "delete_comment_api_v1_comments__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removed."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "The post exists, is visible to you, and isn't yours. Used only where ownership is the question — never to hide a draft, which answers 404."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No such row — or one you aren't allowed to know about. Somebody else's unpublished post answers this, deliberately, because a 403 would confirm that it exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove your comment",
        "tags": [
          "Comments"
        ]
      }
    },
...
    "/api/v1/posts/{post_id}/comments": {
      "get": {
        "description": "One page of a post's conversations, **oldest first**, each with whatever\nwas said back to it.\n\nThe feed runs newest-first because you arrive at it to see what's new. A\ncomment thread is read top to bottom like a conversation, so it runs the\nother way — and the page size is larger, because twenty short lines is\nabout one screen where ten posts is several.\n\n**The page is over top-level comments, and replies come with their\nparent.** Paging the two together would eventually put a page boundary\nthrough the middle of an exchange, which is the one place a boundary must\nnot fall; and a reply can only ever belong to something on the same page as\nitself, so the alternative buys nothing. `total` therefore counts\nconversations — it is what `pages` is computed from — and `total_replies`\nsays how many messages hang off them, so a heading can report the size of\nthe conversation without walking it.",
        "operationId": "get_comments_api_v1_posts__post_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "has_next": false,
                  "has_prev": false,
                  "items": [
                    {
                      "content": "I've walked past that wall a hundred times and never once thought about it.",
                      "created_at": "2026-09-09T07:02:55.104Z",
                      "id": 31,
                      "post_id": 12,
                      "user": {
                        "created_at": "2026-08-30T17:45:11.002Z",
                        "id": 9,
                        "username": "tessaward"
                      },
                      "user_id": 9
                    }
                  ],
                  "page": 1,
                  "page_size": 10,
                  "pages": 1,
                  "total": 5
                },
                "schema": {
                  "$ref": "#/components/schemas/CommentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No such row — or one you aren't allowed to know about. Somebody else's unpublished post answers this, deliberately, because a 403 would confirm that it exists."
          },
          "422": {
            "description": "The body or the query string didn't validate. FastAPI's own shape: `detail` is a list of `{loc, msg, type}`, one per field."
          }
        },
        "summary": "The conversation on a post",
        "tags": [
          "Comments"
        ]
      },
      "post": {
        "description": "Add a comment to a post you can see, or a reply to one of its comments.\n\nWhitespace is trimmed and an empty comment is a 422 — the cap is enforced\nhere and not only in the form, because this is the field a script would\npoint at first.\n\n`parent_id` makes it a reply, and there are exactly two ways that can be\nrefused. A parent that isn't a comment on *this* post is a 404 rather than\na 400: from here it does not exist, and saying otherwise would confirm the\nexistence of comments on a draft somebody hasn't published. A parent that\nis itself a reply is a 400, because the request is well-formed and simply\nasks for something this app does not do.",
        "operationId": "create_comment_api_v1_posts__post_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "content": "I've walked past that wall a hundred times and never once thought about it.",
                  "created_at": "2026-09-09T07:02:55.104Z",
                  "id": 31,
                  "post_id": 12,
                  "user": {
                    "created_at": "2026-08-30T17:45:11.002Z",
                    "id": 9,
                    "username": "tessaward"
                  },
                  "user_id": 9
                },
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "The request was understood and asks for nothing."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No access token, or one that has expired, been tampered with, or belongs to a deleted account. Trade the refresh cookie for a new token at `POST /auth/refresh` rather than sending the person back to a sign-in form."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detail"
                }
              }
            },
            "description": "No such row — or one you aren't allowed to know about. Somebody else's unpublished post answers this, deliberately, because a 403 would confirm that it exists."
          },
          "422": {
            "description": "The body or the query string didn't validate. FastAPI's own shape: `detail` is a list of `{loc, msg, type}`, one per field."
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Say something about a post",
        "tags": [
          "Comments"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A local development server",
      "url": "http://localhost:8000"
    }
  ],
...
}